+++ avm/res/resources/resource-group/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The latest version of the changelog can be found [here](https://github.com/Azure/bicep-registry-modules/blob/main/avm/res/resources/resource-group/CHANGELOG.md).
 
+## 0.4.4
+
+### Changes
+
+- Fixed child module deployment names creation to prevent conflicts when two Resource Groups are deployed in the same subscription at the same time.
+- Updated API version for `Microsoft.Resources/deployments` to `2025-04-01` for telemetry deployment resource.
+
+### Breaking Changes
+
+- None
+
 ## 0.4.3
 
 ### Changes